+++ packages/esm-appointments-app/src/appointment-list-detail/appointment-list-detail.scss
