+++ packages/esm-appointments-app/src/appointments/patient-list-list.scss
@@ -1,4 +1,4 @@
-@import '../../root.scss';
+@import '../root.scss';
 
 title {
   width: 6.938rem;
