+++ packages/esm-appointments-app/src/clinical-appointments.component.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import PatientListList from './appointment-list/patient-list-list/patient-list-list.component';
-import ActiveVisitsTable from './appointment-tabs/booked/active-visits-table.component';
+import PatientListList from './appointments/patient-list-list.component';
+import AppointmentsTable from './tabs/appointments-table.component';
 import ClinicMetrics from './appointments-metrics/appointments-metrics.component';
 
 interface ClinicalAppointmentsProps {}
@@ -9,7 +9,6 @@ const ClinicalAppointments: React.FC<ClinicalAppointmentsProps> = () => {
   return (
     <div>
       <ClinicMetrics />
-      <ActiveVisitsTable />
       <PatientListList />
     </div>
   );
