+++ packages/esm-appointments-app/src/patient-table/patient-table.scss
