+++ packages/esm-appointments-app/src/ui-components/create-edit-patient-list/create-edit-patient-list.scss
@@ -1,4 +1,4 @@
-@import '../../../root.scss';
+@import '../../root.scss';
 
 .header {
   @include carbon--type-style('productive-heading-02');
