+++ packages/esm-appointments-app/src/ui-components/overflow-menu/overflow-menu.scss
